move trusted into script-bindings
